backend/requirements.txt
langgraph>=0.2
langchain-anthropic>=0.3
langchain-community>=0.3
anthropic>=0.39
streamlit>=1.38
chromadb>=0.5
langsmith>=0.1
pandas>=2.2
plotly>=5.22
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0
python-pptx>=1.0
openpyxl>=3.1
